Moved all iTunes converters to their own files
